further refactoring of mutation related code
